refactor: use LiveView hook for Apple Music auth button
Replace broad document.addEventListener click handler with a targeted
phx-hook on the button, and deduplicate the apple_music_developer_token()
call in the navbar template.
